reorganize ui
