+++ example/ui/rag/RagBasic.tsx
